// backend/nest-server/src/types/deezer.ts
export interface DeezerTrack {
  id: number;
  title: string;
  title_short: string;
  title_version: string;
  link: string;
  duration: number;
  rank: number;
  explicit_lyrics: boolean;
  explicit_content_lyrics: number;
  explicit_content_cover: number;
  preview: string;
  md5_image: string;
  time_add: number;
  artist: {
    id: number;
    name: string;
    link: string;
    picture: string;
    picture_small: string;
    picture_medium: string;
    picture_big: string;
    picture_xl: string;
    tracklist: string;
    type: string;
  };
  album: {
    id: number;
    title: string;
    cover: string;
    cover_small: string;
    cover_medium: string;
    cover_big: string;
    cover_xl: string;
    tracklist: string;
    type: string;
  };
  type: string;
}

export interface FormattedTrack {
  id: number;
  title: string;
  artist: string;
  album: string;
  duration: number;
  coverUrl: string;
  previewUrl: string;
  artistId: number;
  albumId: number;
  isExplicit: boolean;
  rank: number;
  isLiked?: boolean;
}

export interface ArtistInfo {
  id: number;
  name: string;
  picture: string;
  tracklist: string;
  type: string;
}